REfs And Some Unknown/Undiscovered Bugs Fixed
Turns out that some commands weren't checking for analysis status before
issuing the actual API endpoint call. This generated incorrect error
messages that can be misleading to non-developers.

I caught the bug by chance and fixed it. Also improved the error
messages to be a bit more helpful ;-)

+++ Source/Plugin.h
@@ -81,7 +81,7 @@ extern "C" {
         RzCore* core,
         CString fcn_name,
         Size    max_results,
-        Float32 confidence,
+        Int32   confidence,
         Bool    debug_mode
     );
